Added entire test suite and a bit in terms of copy constructors/destructors
The stock tests wont run because theres something wrong with the deconstructor. not sure where its happening but it seems like in the waitlist the first and last pointers are going from being null to having empty strings (that aren't allocated to it??) so things go high key wrong when the destructor tries to run? idk can yall take a look ive been staring at it for hours